### 2024-06-24 #6926 @editorialbot
[PRE REVIEW]: Tenta: Remote and Real-Time Sensor Network Management
**Submitting author:** @empicano (<a href="http://orcid.org/0009-0002-2179-9934">Felix Böhm</a>)
**Repository:** https://github.com/iterize/tenta
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
