+++ publication/paper.bib
@@ -6,7 +6,8 @@
   number={2},
   pages={1111--1126},
   year={2021},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  doi={10.5194/amt-14-1111-2021}
 }
 @proceedings{midcost,
   title={Advancing Urban Greenhouse Gas Monitoring: Development and Evaluation of a High-Density CO2 Sensor Network in Munich},
@@ -22,7 +23,8 @@
   number={7},
   pages={3815--3834},
   year={2020},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  doi={10.5194/amt-13-3815-2020}
 }
 @article{zweifel2021trees,
   title={Why trees grow at night},
@@ -32,14 +34,16 @@
   number={6},
   pages={2174--2185},
   year={2021},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/nph.17552}
 }
 @inproceedings{zhang2004hardware,
   title={Hardware design experiences in ZebraNet},
   author={Zhang, Pei and Sadler, Christopher M and Lyon, Stephen A and Martonosi, Margaret},
   booktitle={Proceedings of the 2nd international conference on Embedded networked sensor systems},
   pages={227--238},
-  year={2004}
+  year={2004},
+  doi={10.1145/1031495.1031522}
 }
 @article{werner2006deploying,
   title={Deploying a wireless sensor network on an active volcano},
@@ -49,14 +53,16 @@
   number={2},
   pages={18--25},
   year={2006},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/MIC.2006.26}
 }
 @inproceedings{wenzel2021stand,
   title={Stand-alone low-cost sensor network in the inner city of Munich for modeling urban air pollutants},
   author={Wenzel, Adrian and Chen, Jia and Dietrich, Florian and Thekkekara, Sebastian T and Zollitsch, Daniel and Voggenreiter, Benno and Setili, Luca and Wenig, Mark and Keutsch, Frank N},
   booktitle={EGU General Assembly Conference Abstracts},
   pages={EGU21--15182},
-  year={2021}
+  year={2021},
+  doi={10.5194/egusphere-egu21-15182}
 }
 @software{frost,
   title = {FROST Server},
@@ -111,42 +117,6 @@
   url = {https://github.com/iobridge/thingspeak},
   year = {2011}
 }
-@software{pahoj,
-  title = {paho.mqtt.java},
-  author = {{Eclipse Foundation}},
-  journal = {GitHub repository},
-  publisher = {GitHub},
-  url = {https://github.com/eclipse/paho.mqtt.java},
-  version = {1.2.5},
-  date = {2020-07-15}
-}
-@software{pahoc,
-  title = {paho.mqtt.c},
-  author = {{Eclipse Foundation}},
-  journal = {GitHub repository},
-  publisher = {GitHub},
-  url = {https://github.com/eclipse/paho.mqtt.c},
-  version = {1.3.13},
-  date = {2023-10-19}
-}
-@software{pahop,
-  title = {paho.mqtt.python},
-  author = {{Eclipse Foundation}},
-  journal = {GitHub repository},
-  publisher = {GitHub},
-  url = {https://github.com/eclipse/paho.mqtt.python},
-  version = {2.0.0},
-  date = {2024-02-10}
-}
-@software{mqttjs,
-  title = {MQTT.js},
-  author = {Adam Rudd},
-  journal = {GitHub repository},
-  publisher = {GitHub},
-  url = {https://github.com/mqttjs/MQTT.js},
-  version = {5.5.3},
-  date = {2024-04-19}
-}
[…]
