### 2024-06-27 #6926 @kthyng
@empicano I am going to ping the editorial board to have them see if this is in scope for JOSS, mainly in terms of whether this counts as "research software". This will probably take a few weeks, thanks for your patience.
