### 2024-06-29 #6926 @empicano
Thank you for the update! I hope that I'm allowed to give some input:

Our [research group](https://www.ee.cit.tum.de/en/esm/home/) builds sensor networks that measure environmental data like air quality and CO2. This data can e.g. be used to validate emission inventory data or find previously unknown emission sources.

Tenta allows (1.) to collect measurements of sensor networks remotely. This is not just a "nice-to-have", but a requirement for mid to large sensor networks. The sensor network that Tenta powers at our research group is deployed on rooftops that are difficult to access. We need remote management for such research, as regular physical visits would not be possible. Similarly, Tenta can enable research with sensor networks in remote areas. Tenta allows (2.) to configure sensors remotely. This allows us to react to sensor hardware and software failures much faster and thus increases data yield. Furthermore, this functionality allows running experiments while the sensors are in the field. The fact that Tenta links configurations with measurements leads to highly reproducible datasets.

In our research group Tenta thus ["supports the functioning of research instruments or the execution of research experiments"](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) (from JOSS's documentation: "What we mean by research software"). There is similar research to our CO2 sensor network e.g. cited in the introduction or [ICOS Cities](https://www.icos-cp.eu/projects/icos-cities), where we believe Tenta can be useful as well.

Our research group has previously published related software in JOSS that is also used to support sensor networks with [Pyra: Automated EM27/SUN Greenhouse Gas Measurement Software](https://github.com/openjournals/joss-reviews/issues/5131). I believe that [Thing Directory: Simple and lightweight registry of IoT device metadata](https://github.com/openjournals/joss-reviews/issues/3075) is another JOSS publication that is very similar to this Tenta paper submission.

I also hope that I can alleviate some concerns about the dashboard: The main contribution is the server with its data processing/storage and MQTT communication with the sensors (this also reflects in the amount of work that went into them: The dashboard was developed in around 2 months, while the server took a year of full time work). The server is useful and usable on its own, the dashboard only provides a nice interface to the REST endpoints. In fact, we ran the server without any dashboard for quite a while ourselves in the beginning. The server is well tested and can easily be run locally, as the test and development scripts automatically spin up PostgreSQL/TimescaleDB and Mosquitto instances via Docker.

Thank you for your consideration.
