### 2024-07-01 #6926 @kthyng
@empicano Thank you, yes, the editors that take a look will see your comments to add to their understanding of the software.
