### 2024-07-10 #6926 @kthyng
@faroit Would this be in your wheelhouse at all?
