### 2024-07-10 #6926 @faroit
@kthyng I thoroughly checked the submission but I have to say that, unfortunately, I'm not comfortable editing this. It's neither in my scientific domain nor am I up-to-date in web app dev.

If there is absolutely no one else available I could do it but not starting before mid-August.
