+++ publication/paper.md
@@ -98,6 +98,6 @@ Tenta leverages concurrency to achieve high throughput. Database and HTTP reques
 
 # Author contributions
 
-FB: Designed and developed the software, wrote the tests, wrote the documentation, wrote the manuscript; MM: Developed the software running on the sensors, developed the dashboard, reviewed the manuscript; PA: Supervised the project, extracted the requirements, reviewed the manuscript; JC: Principle investigator, supervised the project, reviewed the manuscript;
+FB: Designed and developed the server software, wrote its tests and documentation, and wrote the manuscript; MM: Developed the software running on the sensors, developed the dashboard, reviewed the manuscript; PA: Supervised the project, extracted the requirements, reviewed the manuscript; JC: Principle investigator, supervised the project, reviewed the manuscript;
 
 # References
