### 2024-07-31 #6926 @dostuffthatmatters
Hi @kthyng, hi @kanishkan91,

Thank you, @empicano, for elaborating on why this counts as research software.

The software is running stable in production on 20 sensor systems even one year after @empicano left our research group.

The dashboard directly offers a suitable User Interface for versionized device configurations. The UIs of competing tools (Thingsboard and equivalent) require workarounds to achieve a similar result but with more clicks and window changes.

I just updated the documentation on how to host the dashboard.

Thank you for editing this paper!
Best regards,

Moritz Makowski (second author of the paper)
Doctoral Candidate
Environmental Sensing and Modeling | Technical University of Munich
