### 2024-08-02 7aef440
Update dashboard build instructions
