### 2024-08-02 0ae19d3
Add `Dockerfile` for dashboard
