+++ .github/workflows/publication.yml
@@ -17,7 +17,7 @@ jobs:
           journal: joss
           paper-path: publication/paper.md
       - name: Upload artifact
-        uses: actions/upload-artifact@v1
+        uses: actions/upload-artifact@v4
         with:
           name: paper
           path: publication/paper.pdf
