### 2024-10-04 7047069
Add JOSS badge to README
