+++ README.md
@@ -16,6 +16,8 @@ _Read the documentation at [tenta.onrender.com](https://tenta.onrender.com/)_
 
 ## More
 
+**Publication:** [![status](https://joss.theoj.org/papers/5daf8d2d13c01da24e949c20a08d29d0/status.svg)](https://joss.theoj.org/papers/5daf8d2d13c01da24e949c20a08d29d0)
+
 **License:** Tenta is licensed under the [MIT License](https://github.com/iterize/tenta/blob/main/LICENSE).
 
 **Research:** We are open for collaborations! If you want to use Tenta in your research, don't hesitate to reach out to contact@iterize.dev. We are happy to help you get started and provide support.
