### 2024-10-09 d2c4dd8
Update author contributions
