### 2024-12-21 #7311 @sash19
Hi, I have reviewed Tenta! My main comments are below. Please feel free to follow up with me for additional reviews or discussions:

- There doesn't seem to be any testing framework explicitly outlined in the documentation to verify the functionality of the software.
- The installation process is described in the documentation, but it can be extremely difficult if needed software (specifically poetry, python, and nodejs) are not installed. Many users may not have python3.11 pre-installed and the software doesn't seem to work with anything else. Same goes with poetry. Installing and correctly setting up the environment is not properly outlined. "Install xxx dependency" is not helpful when the installation process can be lengthy and convoluted. Changing or adding python versions on most systems, for example, requires administrator permissions. However, other software especially package managers like pip and poetry may not work in a sudo/administrator environment as they separately manage user packages.
- The installation issues above can be quite a pain for someone not very savvy or comfortable with such things.
