### 2024-12-22 #7311 @abhishektiwari
@kanishkan91  I have completed my checklist and review.

The software addresses a clear need in sensor network management, particularly for research applications. The architecture is well-thought-out, emphasising flexibility and scalability. Good test coverage for both client and server. I recommend to accept the paper.

## Few languages suggestions

Line 11-12: "when these networks are more than a few sensors in scale and/or their locations difficult to access"

Should be: "when these networks are more than a few sensors in scale and/or their locations are difficult to access"

Line 31-32: The IoT space is dominated by proprietary services e.g. from Amazon and Microsoft Azure

Something like: The IoT space is dominated by proprietary cloud services, e.g., from Amazon Web Services and Microsoft Azure.

Line 89: "CO2 sensor network in munich."

"Munich" should be capitalized

## Nonblocking comment
I know paper mentioned the real-world validation through deployment with 20 CO2 sensors. It will be great if that was shared as working example as part of repository or through a tutorial on documentation pages. This will reduce the friction for anyone trying to figure out use case of this software.

Consider releasing dashboard as NPM package and server as Python module. This is mainly to reduce the friction to use/onboard to Tenta as well as introduce opportunities to extend core functionality via extension and plugins. This may require some change in API interfaces.
