+++ publication/paper.bib
@@ -10,13 +10,13 @@
   doi={10.5194/amt-14-1111-2021}
 }
 @proceedings{midcost,
-  title={Advancing Urban Greenhouse Gas Monitoring: Development and Evaluation of a High-Density CO2 Sensor Network in Munich},
+  title={Advancing Urban Greenhouse Gas Monitoring: Development and Evaluation of a High-Density CO2 Sensor Network in {Munich}},
   author={Aigner, Patrick and K{\"u}bacher, Daniel and Wenzel, Adrian and Schmitt, Adrian and B{\"o}hm, Felix and Makowski, Moritz and Angleitner, Moritz and K{\"u}rzinger, Klaus and Laurent, Olivier and Rubli, Pascal and Grange, Stuart and Emmenegger, Lukas and Chen, Jia},
   booktitle={ICOS Science Conference},
   year={2024},
 }
 @article{muller2020integration,
-  title={Integration and calibration of non-dispersive infrared (NDIR) CO\textsubscript{2} low-cost sensors and their operation in a sensor network covering Switzerland},
+  title={Integration and calibration of non-dispersive infrared (NDIR) CO\textsubscript{2} low-cost sensors and their operation in a sensor network covering {Switzerland}},
   author={M{\"u}ller, Michael and Graf, Peter and Meyer, Jonas and Pentina, Anastasia and Brunner, Dominik and Perez-Cruz, Fernando and H{\"u}glin, Christoph and Emmenegger, Lukas},
   journal={Atmospheric Measurement Techniques},
   volume={13},
@@ -127,7 +127,7 @@
   date = {2024-02-08}
 }
 @article{bart2014high,
-  title={High density ozone monitoring using gas sensitive semi-conductor sensors in the Lower Fraser Valley, British Columbia},
+  title={High density ozone monitoring using gas sensitive semi-conductor sensors in the {Lower Fraser Valley, British Columbia}},
   author={Bart, Mark and Williams, David E and Ainslie, Bruce and McKendry, Ian and Salmond, Jennifer and Grange, Stuart K and Alavi-Shoshtari, Maryam and Steyn, Douw and Henshaw, Geoff S},
   journal={Environmental science \& technology},
   volume={48},
