+++ publication/paper.md
@@ -54,7 +54,7 @@ bibliography: paper.bib
 
 # Summary
 
-In many domains, sensor networks have become indispensable to drive effective decisions: They monitor greenhouse gases and air pollutants [@dietrich2021muccnet; @wenzel2021stand], record volcanic activity [@werner2006deploying], and track animal migrations [@zhang2004hardware]. However, when these networks are more than a few sensors in scale and/or their locations difficult to access, manual inspection of sensors to retrieve their data and check their status gets prohibitively time-intensive. Simultaneously, scale renders the occurrence of sensor failures practically unavoidable. Failures that stay undetected lead to data loss [@bart2014high; @tolle2005design] and skewed results, and thus, potentially, to erroneous conclusions.
+In many domains, sensor networks have become indispensable to drive effective decisions: They monitor greenhouse gases and air pollutants [@dietrich2021muccnet; @wenzel2021stand], record volcanic activity [@werner2006deploying], and track animal migrations [@zhang2004hardware]. However, when these networks are more than a few sensors in scale and/or their locations are difficult to access, manual inspection of sensors to retrieve their data and check their status gets prohibitively time-intensive. Simultaneously, scale renders the occurrence of sensor failures practically unavoidable. Failures that stay undetected lead to data loss [@bart2014high; @tolle2005design] and skewed results, and thus, potentially, to erroneous conclusions.
 
 Tenta is a lightweight server application to manage sensor networks remotely and in real-time. The system receives and stores measurements transmitted by sensors and allows users to monitor sensors via an intuitive dashboard. Additionally, Tenta uses versionized configurations to adapt sensors to changes in requirements. Configurations are schemaless JSON documents that represent a sensor's state. They allow requests to sensors e.g. to update their measurement interval or even their software, as for instance implemented in Ivy [@ivy]. Tenta associates each measurement with the sensor's currently active configuration, which results in highly reproducible datasets. Sensors communicate with Tenta over the widely adopted MQTT protocol. The system exposes a REST API for straightforward task automation. Data is stored in a time series-optimized PostgreSQL+TimescaleDB [@postgres; @timescale] database.
 
@@ -68,7 +68,7 @@ Starting in 2022, our research group deployed a network of 20 mid-precision CO2 
 
 # Statement of need
 
-The IoT space is dominated by proprietary services e.g. from Amazon and Microsoft Azure. These platforms offer a holistic suite of message brokers, monitoring, device management, data storage, and analytics. Several research projects [@muller2020integration; @zweifel2021trees; @burri2019did] use the sensors and infrastructure of the Swiss company Decentlab. Measurements are transmitted via LoRaWAN to gateways that relay them to Decentlab's infrastructure over the Internet. Other projects [@werner2006deploying; @zhang2004hardware] implement custom solutions e.g. based on flooding protocols. Smaller networks sometimes forgo connectivity entirely and instead rely on physical visits to the sites.
+The IoT space is dominated by proprietary cloud services, e.g., from Amazon Web Services and Microsoft Azure. These platforms offer a holistic suite of message brokers, monitoring, device management, data storage, and analytics. Several research projects [@muller2020integration; @zweifel2021trees; @burri2019did] use the sensors and infrastructure of the Swiss company Decentlab. Measurements are transmitted via LoRaWAN to gateways that relay them to Decentlab's infrastructure over the Internet. Other projects [@werner2006deploying; @zhang2004hardware] implement custom solutions e.g. based on flooding protocols. Smaller networks sometimes forgo connectivity entirely and instead rely on physical visits to the sites.
 
[…]
