### 2024-12-27 9c19d24
Add docker compose file for easy experimentation
