### 2024-12-27 #7311 @empicano
Thank you @sash19 and thank you @abhishektiwari for your detailed reviews!

---

@sash19: The tests can be run with the [`./scripts/test`](https://github.com/iterize/tenta/blob/main/server/scripts/test) script, which spins up PostgreSQL and Mosquitto instances and then runs the tests with pytest. This is noted in the [development documentation](https://tenta.onrender.com/contribute). The test coverage is currently at ~94%. Please let me know if I can clarify anything here!

Thanks also for reporting about the difficulties with the setup. I've added a Docker Compose file that handles dependencies like Python 3.11 and poetry automatically, and updated the README with instructions on how to run Tenta with Docker Compose.

---

@abhishektiwari: I adopted your wording suggestions verbatim. We've open-sourced the code of our CO2 sensor network as a real-word example and [linked it in the documentation](https://tenta.onrender.com/community). Interesting idea to release Tenta as NPM package and Python module, too. This would indeed require some API changes, but you're right that the extensibility could be worth it. I will keep this in mind as future improvement.

---

Please let me know if any further changes are needed.
