### 2024-12-28 #7311 @empicano
@kanishkan91, thanks for checking back! I made the changes in commit https://github.com/iterize/tenta/commit/cb3c30ff3816df1f603489318238df2731e1eca0 and https://github.com/iterize/tenta/commit/9c19d245e956c64df4d9b09dae2e512b2aee4390.
