### 2024-12-29 2b1b7a5
Update Ivy citation
