+++ publication/paper.bib
@@ -168,8 +168,12 @@
   doi={10.21105/joss.03075}
 }
 @software{ivy,
-  title = {Ivy},
-  author = {Makowski, Moritz},
-  version = {0.1.0},
-  year = {2024}
+  author       = {Makowski, Moritz and Böhm, Felix and Chen, Jia and Aigner, Patrick},
+  title        = {Ivy: A Data Acquisition System for Distributed Sensor Networks Supporting Remote Configuration and Software Updates},
+  month        = dec,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {1.0.0},
+  doi          = {10.5281/zenodo.14562883},
+  url          = {https://doi.org/10.5281/zenodo.14562883},
 }
