### 2024-12-29 c658774
Update Ivy citation
