+++ publication/paper.bib
@@ -174,6 +174,6 @@
   year         = 2024,
   publisher    = {Zenodo},
   version      = {1.0.0},
-  doi          = {10.5281/zenodo.14562883},
-  url          = {https://doi.org/10.5281/zenodo.14562883},
+  doi          = {10.5281/zenodo.14562882},
+  url          = {https://doi.org/10.5281/zenodo.14562882},
 }
