### 2024-12-29 #7311 @dostuffthatmatters
@abhishektiwari: Regarding the CO2 sensor network mentioned in the paper.

I have updated the citation to [Ivy](https://github.com/tum-esm/ivy) in the paper – we just now got around to make it open-source. It is cited in the Tenta paper as it shows, how Tenta can be used to perform remote software updates. Ivy is based on the software of the network of 20 CO2 sensors in Munich (Hermes).

Both Ivy and Hermes are linked in the community projects.
