### 2025-01-03 #7311 @sash19
Hi, I did review the addition of the docker compose file. This works very well on my system! I like how easy it was to spin up a Tenta instance as compared to the installation process. This method can be added to the documentation or it being in the readme is fine too. I have updated the checklist. Thanks!
