### 2025-01-03 #7311 @empicano
@kanishkan91, thanks for the update! I just released v0.2.0 and archived Tenta with Zenodo. The DOI is https://doi.org/10.5281/zenodo.14597567.
