### 2025-01-05 #7311 @kanishkan91
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ✅ OK DOIs
>
> - 10.5194/amt-14-1111-2021 is OK
> - 10.5194/amt-13-3815-2020 is OK
> - 10.1111/nph.17552 is OK
> - 10.1145/1031495.1031522 is OK
> - 10.1109/MIC.2006.26 is OK
> - 10.5194/egusphere-egu21-15182 is OK
> - 10.1021/es404610t is OK
> - 10.1109/EWSN.2005.1462004 is OK
> - 10.12854/erde-2019-420 is OK
> - 10.21105/joss.03075 is OK
> - 10.5281/zenodo.14562882 is OK
>
> 🟡 SKIP DOIs
>
> - No DOI given, and none found for title: Advancing Urban Greenhouse Gas Monitoring: Develop...
> - No DOI given, and none found for title: FROST Server
> - No DOI given, and none found for title: Grafana
> - No DOI given, and none found for title: TimescaleDB
> - No DOI given, and none found for title: ThingsBoard
> - No DOI given, and none found for title: ThingSpeak
> - No DOI given, and none found for title: PostgreSQL
>
> ❌ MISSING DOIs
>
> - 10.1017/cbo9780511586491.007 may be a valid DOI for title: polars
>
> ❌ INVALID DOIs
>
> - None
> ```

@empicano , @dostuffthatmatters - Could you take a look at the missing DOI in the above?
