### 2025-01-05 #7311 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
