### 2025-01-06 #7311 @empicano
> Could you take a look at the missing DOI in the above?

`10.1017/cbo9780511586491.007` is not the correct DOI for this citation. This cites https://github.com/pola-rs/polars, for which I sadly haven't found a DOI for.
