+++ publication/paper.bib
@@ -71,16 +71,17 @@
   publisher = {GitHub},
   url = {https://github.com/FraunhoferIOSB/FROST-Server},
   version = {2.3.1},
-  date = {2024-03-08},
+  date = {2024-03-08}
 }
 @software{polars,
-  title = {polars},
-  author = {Ritchie Vink},
-  journal = {GitHub repository},
-  publisher = {GitHub},
-  url = {https://github.com/pola-rs/polars},
-  version = {0.20.17},
-  date = {2024-03-28}
+  doi = {10.5281/zenodo.7697217},
+  url = {https://zenodo.org/doi/10.5281/zenodo.7697217},
+  author = {Ritchie Vink,   and Stijn de Gooijer,   and Alexander Beedie,   and Marco Edward Gorelli,   and nameexhaustion,   and Orson Peters,   and Gijs Burghoorn,   and Weijie Guo,   and J van Zundert,   and Gert Hulselmans,   and Cory Grinstead,   and Marshall,   and chielP,   and Itamar Turner-Trauring,   and Lawrence Mitchell,   and Luke Manley,   and Matteo Santamaria,   and Daniël Heres,   and Henry Harbeck,   and Josh Magarick,   and Karl Genockey,   and ibENPC,   and deanm0000,   and Ion Koutsouris,   and Moritz Wilksch,   and eitsupi,   and Jorge Leitao,   and Mick van Gelderen,   and Rodrigo Girão Serrão,  },
+  title = {pola-rs/polars: Python Polars 1.19.0},
+  publisher = {Zenodo},
+  year = {2025},
+  month = {jan},
+  version = {py-1.19.0}
 }
 @software{grafana,
   title = {Grafana},
