### 2025-01-06 #7311 @dostuffthatmatters
I added the correct DOI for Polars.
