### 2025-01-08 #7311 @kthyng
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, "munich" is not capitalized but please check all references carefully.
