### 2025-01-08 53e73e8
Keep country name capitalization in references
