+++ publication/paper.bib
@@ -57,7 +57,7 @@
   doi={10.1109/MIC.2006.26}
 }
 @inproceedings{wenzel2021stand,
-  title={Stand-alone low-cost sensor network in the inner city of Munich for modeling urban air pollutants},
+  title={Stand-alone low-cost sensor network in the inner city of {Munich} for modeling urban air pollutants},
   author={Wenzel, Adrian and Chen, Jia and Dietrich, Florian and Thekkekara, Sebastian T and Zollitsch, Daniel and Voggenreiter, Benno and Setili, Luca and Wenig, Mark and Keutsch, Frank N},
   booktitle={EGU General Assembly Conference Abstracts},
   pages={EGU21--15182},
@@ -148,7 +148,7 @@
   doi={10.1109/EWSN.2005.1462004}
 }
 @article{burri2019did,
-  title={How did Swiss forest trees respond to the hot summer 2015?},
+  title={How did {Swiss} forest trees respond to the hot summer 2015?},
   author={Burri, Susanne and Haeler, Elena and Eugster, Werner and Haeni, Matthias and Etzold, Sophia and Walthert, Lorenz and Braun, Sabine and Zweifel, Roman},
   journal={Die Erde},
   volume={150},
