+++ publication/paper.bib
@@ -10,7 +10,7 @@
   doi={10.5194/amt-14-1111-2021}
 }
 @proceedings{midcost,
-  title={Advancing Urban Greenhouse Gas Monitoring: Development and Evaluation of a High-Density CO2 Sensor Network in {Munich}},
+  title={Advancing Urban Greenhouse Gas Monitoring: Development and Evaluation of a High-Density CO\textsubscript{2} Sensor Network in {Munich}},
   author={Aigner, Patrick and K{\"u}bacher, Daniel and Wenzel, Adrian and Schmitt, Adrian and B{\"o}hm, Felix and Makowski, Moritz and Angleitner, Moritz and K{\"u}rzinger, Klaus and Laurent, Olivier and Rubli, Pascal and Grange, Stuart and Emmenegger, Lukas and Chen, Jia},
   booktitle={ICOS Science Conference},
   year={2024},
