### 2025-01-08 #7311 @empicano
Thank you @kthyng for double checking!

> "munich" is not capitalized

Good catch! I added brackets around "Munich" and "Swiss".
