### 2025-01-09 #7311 @kthyng
Congratulations on your new publication @empicano and @dostuffthatmatters! Many thanks to editor @kanishkan91 and to reviewers @abhishektiwari and @sash19 for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
